Create Stay Service

+++ src/main/java/com/hmy/staybooking/service/StayService.java
@@ -1,14 +1,17 @@
 package com.hmy.staybooking.service;
 
+import com.hmy.staybooking.exception.StayDeleteException;
 import com.hmy.staybooking.exception.StayNotExistException;
-import com.hmy.staybooking.model.Stay;
-import com.hmy.staybooking.model.StayImage;
-import com.hmy.staybooking.model.User;
+import com.hmy.staybooking.model.*;
+import com.hmy.staybooking.repository.LocationRepository;
+import com.hmy.staybooking.repository.ReservationRepository;
 import com.hmy.staybooking.repository.StayRepository;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Isolation;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.multipart.MultipartFile;
 
+import java.time.LocalDate;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
@@ -20,9 +23,18 @@ public class StayService {
 
     private final ImageStorageService imageStorageService;
 
-    public StayService(StayRepository stayRepository, ImageStorageService imageStorageService) {
+    private final LocationRepository locationRepository;
+
+    private final GeoCodingService geoCodingService;
+
+    private final ReservationRepository reservationRepository;
+
+    public StayService(StayRepository stayRepository, LocationRepository locationRepository, ImageStorageService imageStorageService, GeoCodingService geoCodingService, ReservationRepository reservationRepository) {
         this.stayRepository = stayRepository;
+        this.locationRepository = locationRepository;
         this.imageStorageService = imageStorageService;
+        this.geoCodingService = geoCodingService;
+        this.reservationRepository = reservationRepository;
     }
 
     public List<Stay> listByUser(String username) {
@@ -41,22 +53,31 @@ public class StayService {
 
     @Transactional
     public void add(Stay stay, MultipartFile[] images) {
-        List<String> mediaLinks = Arrays.stream(images).parallel().map(image -> imageStorageService.save(image)).collect(Collectors.toList());
+        List<String> mediaLinks = Arrays.stream(images).parallel().map(
+                image -> imageStorageService.save(image)
+        ).collect(Collectors.toList());
+
         List<StayImage> stayImages = new ArrayList<>();
         for (String mediaLink : mediaLinks) {
             stayImages.add(new StayImage(mediaLink, stay));
         }
-        stay.setImages(stayImages);
 
+        stay.setImages(stayImages);
         stayRepository.save(stay);
+
+        Location location = geoCodingService.getLatLng(stay.getId(), stay.getAddress());
+        locationRepository.save(location);
     }
 
-    @Transactional
-    public void delete(Long stayId, String username) {
-        User user = new User.Builder().setUsername(username).build();
-        Stay stay = stayRepository.findByIdAndHost(stayId, user);
+    @Transactional(isolation = Isolation.SERIALIZABLE)
+    public void delete(Long stayId, String username) throws StayNotExistException, StayDeleteException {
+        Stay stay = stayRepository.findByIdAndHost(stayId, new User.Builder().setUsername(username).build());
         if (stay == null) {
             throw new StayNotExistException("Stay doesn't exist");
+        }
+        List<Reservation> reservations = reservationRepository.findByStayAndCheckoutDateAfter(stay, LocalDate.now());
+        if (reservations != null && reservations.size() > 0) {
+            throw new StayDeleteException("Cannot delete stay with active reservation");
         }
 
         stayRepository.deleteById(stayId);
